requirements.txt
#backend
fastapi
uvicorn
pydantic
sqlalchemy
alembic
requests
Pillow  # Aggiunto per la manipolazione delle immagini
pandas
passlib
python-jose
python-multipart
bcrypt

#frontend
streamlit

